requirements.txt
bleak==0.20.2
bleak-retry-connector==3.1.1
pycryptodome==3.18.0